use a single StreamProcessorMock definition
